requirements.txt
# requirements.txt mis à jour
customtkinter>=5.2.2
openai-whisper>=20240930
yt-dlp>=2024.8.0
requests>=2.31.0
anthropic>=0.70.0
openai>=1.50.0
google-genai>=1.0.0
Pillow>=11.0.0
torch>=2.4.0
torchaudio>=2.4.0